// src/api-client/models/asset-response-dto.ts
/* tslint:disable */

/**
 * API Documentation
 * API - IT Asset Tracker
 *
 * The version of the OpenAPI document: 0.0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface AssetResponseDto
 */
export interface AssetResponseDto {
  /**
   * Unique asset identifier
   * @type {number}
   * @memberof AssetResponseDto
   */
  id: number;
  /**
   * Unique stock keeping unit (SKU) code for the asset
   * @type {string}
   * @memberof AssetResponseDto
   */
  sku: string;
  /**
   * Name of the asset
   * @type {string}
   * @memberof AssetResponseDto
   */
  name: string;
  /**
   * Description of the asset
   * @type {string}
   * @memberof AssetResponseDto
   */
  description: string;
  /**
   * Model of the asset
   * @type {string}
   * @memberof AssetResponseDto
   */
  model: string;
  /**
   * Brand of the asset
   * @type {string}
   * @memberof AssetResponseDto
   */
  brand: string;
  /**
   * Category identifier (foreign key)
   * @type {number}
   * @memberof AssetResponseDto
   */
  categoryId: number;
  /**
   * Employee identifier (foreign key) - assigned custodian
   * @type {number}
   * @memberof AssetResponseDto
   */
  employeeId: number;
  /**
   * When the record was created
   * @type {string}
   * @memberof AssetResponseDto
   */
  createdAt: string;
  /**
   * Soft-delete timestamp; null while the record is active
   * @type {string}
   * @memberof AssetResponseDto
   */
  deletedAt: string | null;
  /**
   *
   * @type {string}
   * @memberof AssetResponseDto
   */
  categoryName: string;
  /**
   *
   * @type {string}
   * @memberof AssetResponseDto
   */
  employeeName: string;
}